backend/quiz_service/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pytest==9.0.2
pytest-cov==7.1.0
pytest-asyncio==1.3.0
httpx==0.28.1
grpcio==1.80.0
grpcio-tools==1.80.0
PyJWT[crypto]==2.12.1
pymongo==4.16.0